promtail/default-values.yaml: add wildcard toleration
It is a common pattern to apply taints to cluster node pools in order to
segment workloads. In the vast majority of cases, users will not want
that same segmentation to apply to log collection. This change ensures
that by default logs will always be collected on all nodes, regardless
of taint-based segmentation.